requirements.txt
pyodbc
Flask
SQLAlchemy
Flask-Alembic
uvicorn[standard]
pydantic
azure-identity
lxml